+++ src/components/standardForm/StandardFormField.tsx
@@ -1,6 +1,19 @@
 import React from 'react'
 import classes from './StandardFormField.module.css'
 
-export function StandardFormField({ children }: { children: React.ReactNode }) {
-    return <div className={classes.standardFormField}>{children}</div>
+export function StandardFormField({
+    children,
+    dataTest,
+}: {
+    children: React.ReactNode
+    dataTest?: string
+}) {
+    return (
+        <div
+            className={classes.standardFormField}
+            data-test={dataTest ?? 'standard-form-field'}
+        >
+            {children}
+        </div>
+    )
 }